requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Database
pymysql==1.1.0
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# HTTP Requests
requests==2.31.0

# Date/Time
python-dateutil==2.8.2

# File Operations
pathlib2==2.3.7

# Configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Async Support
aiohttp>=3.11.13
asyncio==3.4.3

# Stock Data APIs
tushare==1.2.89
akshare==1.17.26

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 